🗺️ Atlas: [architectural change] Resolve circular dependencies in storage and core
